Change to using symlink for rename
